+++ packages/terra-responsive-element/CHANGELOG.md
@@ -3,6 +3,18 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* updated package.json test scripts
+
+5.5.0 - (July 30, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+5.4.0 - (July 24, 2019)
+------------------
+### Changed
+* Minor dependency version bump
 
 5.3.0 - (July 16, 2019)
 ------------------